Refactor to use ES6
