+++ server/controllers/announcements.js
@@ -1,10 +1,10 @@
-import { Announcement } from '../models';
+import { Announcement } from '../models/Announcement.js';
 
 /*
 * Returns all announcements for the course found
 * based on the courseId. Returns an empty array if the course
 * is not found.
 */
-exports.findAllByCourse = (courseId) => {
+export function findAllByCourse(courseId) {
 
-};
+}
