+++ server/controllers/assignments.js
@@ -1,8 +1,8 @@
-import { Assignment } from '../models';
+import { Assignment } from '../models/Assignment.js';
 
 /*
 * Returns an array of all assignments for a given course.
 */
-exports.findAllByCourse = (courseId) => {
+export function findAllByCourse(courseId) {
 
-};
+}
