+++ server/controllers/courses.js
@@ -1,17 +1,17 @@
-import { Course } from '../models';
+import { Course } from '../models/Course.js';
 
 /*
 * Returns all courses that the student is enrolled in.
 * Returns an empty array if the student is not found.
 */
-exports.findAllByStudent = (studentId) => {
+export function findAllByStudent(studentId) {
 
-};
+}
 
 /*
 * Returns all courses that the teacher is teaching.
 * Returns an empty array if the teacher is not found.
 */
-exports.findAllByTeacher = (teacherId) => {
+export function findAllByTeacher(teacherId) {
 
-};
+}
