+++ server/controllers/students.js
@@ -1,9 +1,9 @@
-import { Student } from '../models';
+import { Student } from '../models/Student.js';
 
 /*
 * Returns a student object matching the provided id. Returns undefined if the
 * student doesn't exist.
 */
-exports.findById = (studentId) => {
+export function findById(studentId) {
 
-};
+}
