+++ server/controllers/submissions.js
@@ -1,17 +1,17 @@
-import { Submission } from '../models';
+import { Submission } from '../models/Submission.js';
 
 /*
 * Returns all submissions across all courses that the students is enrolled
 * in. Returns an empty array if the student does not exist.
 */
-exports.getAllByStudent = (studentId) => {
+export function getAllByStudent(studentId) {
 
-};
+}
 
 /*
 * Returns an array of all submissions for a given course. Returns an empty
 * array if the course is not found. This should be used for teachers only.
 */
-exports.getAllByCourse = (courseId) => {
+export function getAllByCourse(courseId) {
 
-};
+}
